requirements.txt
appdirs==1.4.4
appnope==0.1.2
astroid==2.4.0
backcall==0.2.0
bidict==0.21.2
black==21.6b0
blinker==1.4
certifi==2020.4.5.1
click==7.1.2
decorator==5.0.9
dill==0.3.4
dnspython==1.16.0
dominate==2.5.1
enum-compat==0.0.3
enum34==1.1.10
eventlet==0.25.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==4.3.1
greenlet==0.4.15
gunicorn==20.0.4
ipython==7.25.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
logilab-common==1.6.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
monotonic==1.5
mypy-extensions==0.4.3
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pydantic==1.8.2
pyfunctional==1.4.3
Pygments==2.9.0
pylint==2.5.0
python-dotenv==0.13.0
python-engineio==3.13.2
python-socketio==4.6.0
regex==2021.7.6
sentry-sdk==0.14.3
six==1.14.0
tabulate==0.8.9
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.2.5
WebOb==1.8.6
Werkzeug==1.0.1
wrapt==1.12.1
